refactor: search-page 관련 도메인 폴더구조 정리
