+++ apps/client/src/pages/search/components/search-home/popular-search-section.css.ts
