+++ apps/client/src/pages/search/components/search-home/recent-festivals-section.css.ts
