+++ apps/client/src/pages/search/components/search-home/recent-search-section.css.ts
