+++ apps/client/src/pages/search/components/search-result/artist/artist-info.tsx
