+++ apps/client/src/pages/search/components/search-result/artist/artist-not-found.css.ts
