+++ apps/client/src/pages/search/components/search-result/artist/artist-not-found.tsx
@@ -1,6 +1,6 @@
 import { IcExclamationCircleOutlined32 } from '@confeti/design-system/icons';
 
-import { SEARCH_NOT_FOUND_MESSAGE } from '../constants/notice-message';
+import { SEARCH_NOT_FOUND_MESSAGE } from '../../../constants/notice-message';
 
 import * as styles from './artist-not-found.css';
