+++ apps/client/src/pages/search/components/search-result/artist/artist-section.css.ts
