+++ apps/client/src/pages/search/components/search-result/common/notice-section.css.ts
