+++ apps/client/src/pages/search/components/search-result/common/title.css.ts
