+++ apps/client/src/pages/search/components/search-result/performance/performance-count-section.tsx
