+++ apps/client/src/pages/search/components/search-result/performance/performance-info.tsx
