+++ apps/client/src/pages/search/components/search-result/performance/performance-section.css.ts
